Remove snyk build wrapper and move installations into advanced section
