--- src/main/resources/io/snyk/jenkins/tools/SnykBuildWrapper/config.jelly
@@ -1,18 +1,0 @@
-<?jelly escape-by-default="true"?>
-<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
-  <j:choose>
-    <j:when test="${descriptor.hasInstallationsAvailable()}">
-      <f:entry title="Snyk version">
-        <j:set var="tools" value="${descriptor.installations}"/>
-        <select name="_.snykVersion" class="setting-input">
-          <j:forEach var="tool" items="${tools}">
-            <f:option value="${tool.name}" selected="${tool.name == instance.snykInstallation}">${tool.name}</f:option>
-          </j:forEach>
-        </select>
-      </f:entry>
-    </j:when>
-    <j:otherwise>
-      <f:entry title="No setup will be done, as no Snyk installations have been defined in the Jenkins system config"/>
-    </j:otherwise>
-  </j:choose>
-</j:jelly>
