--- src/main/resources/io/snyk/jenkins/tools/SnykBuildWrapper/help.html
@@ -1,11 +1,0 @@
-<div>
-  <p>
-    Ensures that the selected version of Snyk tools are installed. In addition,
-    the Snyk tools will be added at the start of the <code>PATH</code> environment
-    variable during builds.
-  </p>
-  <p>
-    If no Snyk installations have been defined in the Jenkins system config, then
-    none of the above steps will take place.
-  </p>
-</div>
